+++ contracts/utils/Arrays.sol
@@ -45,4 +45,66 @@ library Arrays {
             return low;
         }
     }
+
+    /**
+     * @dev Sorts `array` of integers in an ascending order.
+     *
+     * Sorting is done in-place using the heap sort algorithm.
+     * Examples of gas cost with optimizer enabled for 200 runs:
+     * - 10 random items: ~11K gas
+     * - 100 random items: ~224K gas
+     */
+    function sort(uint256[] memory array) internal pure {
+        unchecked {
+            uint256 length = array.length;
+            if (length < 2) return;
+            // Heapify the array
+            for (uint256 i = length / 2; i-- > 0; ) {
+                _siftDown(array, length, i, array[i]);
+            }
+            // Drain all elements from highest to lowest and put them at the end of the array
+            while (--length != 0) {
+                uint256 val = array[0];
+                _siftDown(array, length, 0, array[length]);
+                array[length] = val;
+            }
+        }
+    }
+
+    /**
+     * @dev Insert a `inserted` value into an empty space in a binary heap.
+     * Makes sure that the space and all items below it still form a valid heap.
+     * Index `empty` is considered empty and will be overwritten.
+     */
+    function _siftDown(
+        uint256[] memory array,
+        uint256 length,
+        uint256 empty,
+        uint256 inserted
+    ) private pure {
+        unchecked {
+            while (true) {
+                // The first child of empty, one level deeper in the heap
+                uint256 child = empty * 2 + 1;
+                // Empty has no children
+                if (child >= length) break;
+                uint256 childVal = array[child];
+                uint256 otherChild = child + 1;
+                // Pick the larger child
+                if (otherChild < length) {
+                    uint256 otherChildVal = array[otherChild];
+                    if (otherChildVal > childVal) {
+                        child = otherChild;
+                        childVal = otherChildVal;
+                    }
+                }
+                // No child is larger than the inserted value
+                if (childVal <= inserted) break;
+                // Move the larger child one level up and keep sifting down
+                array[empty] = childVal;
+                empty = child;
+            }
+            array[empty] = inserted;
+        }
+    }
 }